refactor(storage): split remote.rs and git_notes.rs into module directories
Both files exceeded the 400-line QA limit (remote.rs: 582 lines,
git_notes.rs: 421 lines). Converts each to a module directory with
logical sub-files:

- remote/ — mod.rs (struct + MemoryBackend impl), wire_types.rs (HTTP
  request/response types), tests.rs (URL encoding + search wire tests)
- git_notes/ — mod.rs (struct + helpers + append_to_git_notes),
  backend_impl.rs (MemoryBackend trait impl)

No public API changes. All 68 spelunk-core tests pass.

Closes spelunk-cloud/spelunk#366
